requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
moviepy==2.1.1
ngrok==1.4.0
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.15.0
six==1.17.0
tqdm==4.67.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.1
Werkzeug==3.1.3
